Extract date from probation revoked events
Instead of doing a full text fuzzy search, we now do a fuzzy search per event for "probation revoked". We preserve previous behavior by converting the corresponding date string back to a boolean value. Eventually we want to pass through the date of the probation revoked event so that we can use this information during expunger analysis.